Version 1.0.2025.204
Simple program to generate new version numbers.  Format output is $variablename = value.   Takes values for the following:
Major version number:
Minor version number:
Year of build:
Day of year for build:

If major and/or minor versions are omitted then the end user is prompted for those values.  The current system date information is used to fill in the Year of build and day of year.  The output produces two formats, a single string for the complete version and build number and a set of strings containing each element of the version and build number.  This provides the end user with two ways to use values.

Produces the following output:
ps1> .\Gen-ProgramVersion.ps1 -Version 3 -MinorVersion 3
$Version = 3.3.2025.204
$Version = 3
$MinorVersion = 3
$MajorBuild = 2025
$MinorBuild = 204