Version 1.0
The script reads the directory file list storing the full path name to the file, the creation time, last write time, and last access time.  This information is then logged to the console screen.  The user may pass any accessible file/directory path, and may request a recursive read of the directory structure.  The default settings are a non-recursive logging of the current working directory.