ApexOneLogCollector.ps1 -  Version: 3.4.25337
Added function to measure the agent installation directory and write the output to the root of the data collection folder.  The new functions measure the agent install directory for the number of objects, and the sum of the size of those objects.  Each directory level is measured individually.  The top level is not the sum of all sub-directories; only of the objects within the level being measured.